+++ packages/cli/src/bin.ts
@@ -1,4 +1,4 @@
-#!/usr/bin/env node
+#!/usr/bin/env bun
 
 import { runCli } from "./index.js";